# attendance-manager/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AttendanceResponse": {
        "type": "object",
        "properties": {
          "attendanceId": {
            "type": "string",
            "description": "Attendance Id"
          },
          "userId": {
            "type": "string",
            "description": "User Id"
          },
          "meetingId": {
            "type": "string",
            "description": "Meeting Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "UNEXCUSED_ABSENCE",
              "EXCUSED_ABSENCE",
              "PENDING"
            ],
            "description": "Status"
          }
        },
        "required": [
          "attendanceId",
          "userId",
          "meetingId",
          "status"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/attendance": {
      "post": {
        "operationId": "post-attendance",
        "summary": "Posts an Attendance",
        "description": "Creates a new Attendance Record with params",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string",
              "description": "User Id"
            },
            "required": true,
            "description": "User Id",
            "example": "123"
          },
          {
            "in": "path",
            "name": "meetingId",
            "schema": {
              "type": "string",
              "description": "Meeting Id"
            },
            "required": true,
            "description": "Meeting Id",
            "example": "123"
          },
          {
            "in": "path",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PRESENT",
                "UNEXCUSED_ABSENCE",
                "EXCUSED_ABSENCE",
                "PENDING"
              ],
              "description": "Status"
            },
            "required": true,
            "description": "Status",
            "example": "example"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/attendance/{attendanceId}": {
      "patch": {
        "operationId": "patch-attendance-{attendanceId}",
        "summary": "Updates an Attendance",
        "description": "Updates an Attendance Record by attendanceId",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "attendanceId",
            "schema": {
              "type": "string",
              "description": "Attendance Id"
            },
            "required": true,
            "description": "Attendance Id",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-attendance-{attendanceId}",
        "summary": "Deletes an Attendance",
        "description": "Deletes an Attendance Record by attendanceId",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "attendanceId",
            "schema": {
              "type": "string",
              "description": "Attendance Id"
            },
            "required": true,
            "description": "Attendance Id",
            "example": "123"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/attendance/status": {
      "put": {
        "operationId": "put-attendance-status",
        "summary": "",
        "description": "",
        "tags": [
          "Attendance"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}